prevent a race condition + minor fixes and changes
Move the LAN rules to be after uplink rules, prevent more than one instance of the script from running at the same time (at the start of the script) to avoid potential race conditions. We also now check for the `flock` command as a result. 

Also, the github repo and license are included within the script.